requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile
#
build==1.4.0
    # via pip-tools
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via pynacl
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
    #   qrcode
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
packaging==26.0
    # via
    #   build
    #   pytest
    #   wheel
pip-tools==7.5.3
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
pycparser==3.0
    # via cffi
pygments==2.19.2
    # via pytest
pynacl==1.6.2
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.2
    # via -r requirements.in
pyyaml==6.0.3
    # via -r requirements.in
qrcode==8.2
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
urllib3==2.6.3
    # via requests
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
